gallan: pull BatchingUploadHandler into it's own class.
 - with CurrentBatchWriter being extracted, it used less and less of S3Archiver's state, become a prime candidate for pruning.
